workers/requirements.txt
tweepy
flask
jinja2
elasticsearch
requests_aws4auth
boto3
kafka-python